requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==4.2.13
django-grappelli==4.0.1
djangorestframework==3.15.2
idna==3.7
pillow==10.3.0; python_version < "3.13"
pillow>=11.0.0,<12.0.0; python_version >= "3.13"
psycopg2-binary==2.9.9; python_version < "3.13"
psycopg2-binary>=2.9.10,<3.0.0; python_version >= "3.13"
pyTelegramBotAPI==4.22.0
python-dotenv==1.0.1
python-logstash==0.4.8
requests==2.32.3
sqlparse==0.5.0
telebot==0.0.5
typing_extensions==4.12.2
urllib3==1.26.16
asgiref==3.8.1
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2024.7.4
charset-normalizer==3.3.2
Django==4.2.13
django-grappelli==4.0.1
djangorestframework==3.15.2
google==3.0.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
gspread==6.2.1
idna==3.7
oauthlib==3.2.2
pillow==10.3.0; python_version < "3.13"
pillow>=11.0.0,<12.0.0; python_version >= "3.13"
psycopg2-binary==2.9.9; python_version < "3.13"
psycopg2-binary>=2.9.10,<3.0.0; python_version >= "3.13"
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyTelegramBotAPI==4.22.0
python-dotenv==1.0.1
python-logstash==0.4.8
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
soupsieve==2.7
sqlparse==0.5.0
telebot==0.0.5
typing_extensions==4.12.2
urllib3==1.26.16
